Comments for Programming
I really like your code! It is similar to what I did

I added a comment under Initiation cell mentioning that it is for import of packages needed for the data analysis that was missing.

You mention in the code that you use the (* - wildcard) to find files with the .nii, however in the code itself you use the glob.glob function but then you call specifically the “anatomical.nii” and not use the (*) therefore it is still hardpathing (I think).

I edited it in a way not for it to be a more flexible with the wildcard -> e.g. *anatomical*.nii*

You also had a typo in the Visualizing the functional data so I fixed that as well as added spaces where they were missing.